pyproject.toml
[project]
name = "self-hosted-ai"
version = "1.0.0"
description = "Self-hosted AI Kubernetes platform"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Self-Hosted AI Team"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "kubernetes>=28.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "venv",
    "__pycache__",
    "build",
    "dist",
    "htmlcov",
]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
    "SIM",  # flake8-simplify
    "N",    # pep8-naming
    "S",    # flake8-bandit (security)
]
ignore = [
    "E501",  # line length handled by formatter
    "B008",  # do not perform function calls in argument defaults
    "S101",  # use of assert detected
    "S104",  # possible binding to all interfaces
    "S108",  # probable insecure usage of temp file/directory
]

[tool.ruff.lint.isort]
known-first-party = ["agents", "self_hosted_ai"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*.py" = ["S101", "S105", "S106", "ARG"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

[tool.bandit]
exclude_dirs = ["tests", ".venv", "build", "dist"]
skips = ["B101", "B601"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --tb=short"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit: Unit tests (no cluster required)",
    "platform: Cluster infrastructure validation",
    "api: API endpoint tests",
    "integration: Service integration tests",
    "security: Security validation tests",
    "performance: Performance benchmark tests",
    "e2e: End-to-end workflow tests",
    "gpu: Tests requiring GPU access",
    "slow: Tests taking >30 seconds",
    "critical: Tests for deployment readiness gate",
    "asyncio: Tests using async/await",
]
filterwarnings = [
    "ignore::urllib3.exceptions.InsecureRequestWarning",
]

[tool.coverage.run]
source = ["agents"]
omit = ["*/tests/*", "*/test_*.py", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.setuptools]
packages = ["agents"]

[tool.setuptools.package-data]
agents = ["py.typed"]

[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"
